refactor: 패키지 리팩토링
- http패키지 하위 클래스를 was패키지로 이동
- Handler 관련 클래스를 web패키지로 이동
