+++ src/main/java/codesquad/was/http/RequestLine.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.was.http;
 
 public class RequestLine {
