+++ src/main/java/codesquad/was/http/type/HeaderType.java
@@ -1,4 +1,4 @@
-package codesquad.http.type;
+package codesquad.was.http.type;
 
 import java.util.Arrays;
