+++ src/main/java/codesquad/was/http/type/MimeType.java
@@ -1,4 +1,4 @@
-package codesquad.http.type;
+package codesquad.was.http.type;
 
 import java.util.Arrays;
 import java.util.Set;
