+++ src/main/java/codesquad/web/SignUpRequestHandler.java
@@ -1,8 +1,8 @@
 package codesquad.web;
 
-import codesquad.http.HttpRequest;
-import codesquad.http.HttpResponse;
 import codesquad.model.User;
+import codesquad.was.http.HttpRequest;
+import codesquad.was.http.HttpResponse;
 import java.io.IOException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
