+++ src/test/java/codesquad/was/http/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.was.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
