+++ src/test/java/codesquad/was/http/HttpResponseTest.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.was.http;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
